@@ -1,13 +1,15 @@
 apply plugin: 'com.android.application'
+apply plugin: 'org.jetbrains.kotlin.android'
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion "25.0.2"
 
+    compileSdk 34
     defaultConfig {
         applicationId "com.bdt"
-        minSdkVersion 14
-        targetSdkVersion 14
+        minSdkVersion 21
+        targetSdkVersion 34
+        versionCode 6
+        versionName "1.3.3"
     }
 
     buildTypes {
@@ -16,11 +18,14 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.txt'
         }
     }
+    namespace 'com.bdt'
 }
 
 dependencies {
-    compile 'com.android.support:appcompat-v7:25.3.1'
-    compile 'com.android.support:design:25.3.1'
-    compile 'com.android.support:preference-v7:25.3.1';
-    compile 'com.android.support:preference-v14:25.3.1';
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'com.google.android.material:material:1.0.0'
+    implementation 'androidx.preference:preference:1.0.0'
+    implementation 'androidx.legacy:legacy-preference-v14:1.0.0'
+    implementation 'androidx.core:core-ktx:+'
 }
