@@ -1,17 +1,15 @@
 package com.bdt.settings;
 
 import android.os.Bundle;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentTransaction;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.widget.FrameLayout;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentTransaction;
 
 import com.bdt.R;
 
 public class SettingsActivity extends AppCompatActivity {
-
-    FrameLayout mContainer;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
